pkg/usecases/requirements.txt
pandas==2.2.3
flask==3.1.0
flask-restx==1.3.0
flask-cors==5.0.0
paramiko==2.9.3
pytest==8.3.4
pytest-html==4.1.1
pytest-cov==6.0.0
pep8==1.7.1
datetime==4.3
PyYAML==6.0
requests==2.32.2